refactor code base, conform to clean architecture paradigm
