+++ backend/src/config/db.config.ts
@@ -1,0 +1,4 @@
+/**
+ * connect to the DBMS and export the connection for external use
+ *
+ */
