+++ backend/src/index.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import dotenv from 'dotenv';
 import bodyParser from "body-parser";
 import cors from "cors";
-import sensorsRouter from "./routes/sensor.routes";
+import sensorsRouter from "./routes/readings.routes";
 
 dotenv.config();
 const app = express();
