+++ backend/src/persistence/interfaces/readings.persistence.ts
@@ -1,0 +1,6 @@
+/**
+ * Connection to database that returns all sensor readings for the current minute
+ */
+export interface IReadImmediateReadings {
+    (): any[]
+}
