--- backend/src/routes/sensor.routes.ts
@@ -1,8 +1,0 @@
-import express from "express";
-import {streamTimeseries} from '../controllers/timeseries.controllers';
-
-const router = express.Router();
-
-router.get('/stream', streamTimeseries);
-
-export default router;
